+++ src/main/java/com/sparta/nalda/controller/UserController.java
@@ -4,7 +4,6 @@ import com.sparta.nalda.common.MessageResponse;
 import com.sparta.nalda.dto.user.PasswordUpdateRequestDto;
 import com.sparta.nalda.dto.user.SignupRequestDto;
 import com.sparta.nalda.dto.user.UserResponseDto;
-import com.sparta.nalda.service.user.AuthService;
 import com.sparta.nalda.service.user.UserService;
 import com.sparta.nalda.util.AuthUser;
 import jakarta.validation.Valid;
@@ -16,7 +15,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -26,11 +24,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
-    private final AuthService authService;
 
     @PostMapping("/signup")
     public ResponseEntity<MessageResponse> signup(@RequestBody @Valid SignupRequestDto dto) {
-        authService.signup(dto.getEmail(), dto.getPassword(), dto.getAddress(), dto.getUserRole());
+        userService.signup(dto.getEmail(), dto.getPassword(), dto.getAddress(), dto.getUserRole());
         return ResponseEntity.ok(new MessageResponse("회원가입에 성공하였습니다."));
     }
