+++ src/main/java/com/sparta/nalda/service/user/UserService.java
@@ -1,6 +1,7 @@
 package com.sparta.nalda.service.user;
 
 import com.sparta.nalda.dto.user.UserResponseDto;
+import com.sparta.nalda.util.UserRole;
 
 public interface UserService {
 
@@ -11,5 +12,7 @@ public interface UserService {
     void updateUserPassword(Long userId, String oldPassword, String newPassword);
 
     void deleteUser(Long userId, String password);
+
+    void signup(String email, String password, String address, UserRole userRole);
 
 }
