+++ src/main/java/com/sparta/nalda/service/user/UserServiceImpl.java
@@ -5,6 +5,7 @@ import com.sparta.nalda.entity.UserEntity;
 import com.sparta.nalda.entity.WithDrawnEmailEntity;
 import com.sparta.nalda.repository.UserRepository;
 import com.sparta.nalda.repository.WithDrawnEmailRepository;
+import com.sparta.nalda.util.UserRole;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -17,6 +18,22 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final WithDrawnEmailRepository withDrawnEmailRepository;
     private final PasswordEncoder passwordEncoder;
+
+    @Override
+    public void signup(String email, String password, String address, UserRole userRole) {
+        if(userRepository.existsByEmail(email)){
+            throw new IllegalArgumentException("이미 존재하는 이메일입니다.");
+        }
+
+        if(withDrawnEmailRepository.existsByEmail(email)) {
+            throw new IllegalArgumentException("탈퇴한 회원 이메일입니다.");
+        }
+
+        String hashPassword = passwordEncoder.encode(password);
+        UserEntity user = new UserEntity(email, hashPassword, address, userRole);
+        userRepository.save(user);
+
+    }
 
     @Override
     public UserResponseDto getUser(Long userId) {
@@ -61,5 +78,7 @@ public class UserServiceImpl implements UserService {
         userRepository.delete(user);
 
     }
+
+
 
 }
